Renaming qpy directory to quantpy
