+++ quantpy/moving_average.py
